requirements.txt
# Orchestration & Agents
langchain
langchain-openai
langgraph
langserve[all]

# Data Processing
pypdf
langchain-community
tiktoken

# Vector Store
faiss-cpu

# Web Scraping
beautifulsoup4

# Web & API
fastapi
uvicorn
python-dotenv

# Search Tools
tavily-python